[WIP] earthEngine: Add integration for earth engine API Landsat data
This commit adds a helper to fetch landsat data from earth engine using a list of corrdinates.
This will fetch Landsat data from earth engine and return a map for further analysis.

Note: This feature is a Work-In-Progress. This helper is a no-op until the developers get access to earth engine resources.